refactor: move all structure-defining types and structures to the src/Structure folder
