+++ src/Structure/Versions.ts
@@ -1,0 +1,6 @@
+export enum Version {
+  UNKNOWN,
+  V0,
+  V1,
+  V2,
+}
