+++ src/TaskViews/Task.tsx
@@ -2,7 +2,7 @@ import { useState, useRef, useEffect, memo } from "react";
 import PropTypes from "prop-types";
 import TaskContent from "../Tasks/TaskContent.js";
 import React from "react";
-import { FinishedState } from "../Tasks/TaskInterfaces.tsx";
+import { FinishedState } from "../Structure/TaskInterfaces.tsx";
 import {
   translateLastCompletedDatetime,
   translateTimeDifference,
