+++ src/TaskViews/TaskList.js
@@ -5,7 +5,7 @@ import {
   sanitizeTopicOrderIndex,
 } from "../ADG/ModifyFuncGeneratorsV1.ts";
 import { getUnfoldAll, getFoldAll } from "../Topics/TopicModifyFuncGens.js";
-import { FinishedState } from "../Tasks/TaskInterfaces.tsx";
+import { FinishedState } from "../Structure/TaskInterfaces.tsx";
 import { isTaskDueIn, convertDueDateNameToSeconds } from "../Timing.ts";
 import TopicContainer from "../Topics/TopicContainer.tsx";
 import TaskContainer from "./TaskContainer.tsx";
